# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Central management server for the OpenLabStats fleet. Manages agent enrollment, software mappings, lab groupings, and reporting.",
        "title": "OpenLabStats Server API",
        "contact": {},
        "version": "0.1.3"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/agents": {
            "get": {
                "description": "Returns all agents that have registered with the server.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "agents"
                ],
                "summary": "List all enrolled agents",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/store.Agent"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/agents/register": {
            "post": {
                "description": "Agents call this on startup and periodically to register themselves with the server.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "agents"
                ],
                "summary": "Register or heartbeat an agent",
                "parameters": [
                    {
                        "description": "Agent registration payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.RegisterAgentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/store.Agent"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/agents/{agentID}": {
            "get": {
                "description": "Returns a single agent's details.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "agents"
                ],
                "summary": "Get agent by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Agent ID",
                        "name": "agentID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/store.Agent"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Removes an agent from the fleet inventory.",
                "tags": [
                    "agents"
                ],
                "summary": "Remove an agent",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Agent ID",
                        "name": "agentID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/agents/{agentID}/force-update": {
            "post": {
                "description": "Queues an update for the specified agent which will be delivered on its next heartbeat.",
                "tags": [
                    "agents"
                ],
                "summary": "Force an agent to update immediately",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Agent ID",
                        "name": "agentID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/agents/{agentID}/lab": {
            "put": {
                "description": "Associates an agent with a lab/room for grouping.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "agents"
                ],
                "summary": "Assign agent to a lab",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Agent ID",
                        "name": "agentID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Lab assignment",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.AssignLabRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.AssignLabRequest": {
            "type": "object",
            "properties": {
                "labId": {
                    "type": "string"
                }
            }
        },
...
        "api.RegisterAgentRequest": {
            "type": "object",
            "properties": {
                "agentVersion": {
                    "type": "string"
                },
                "building": {
                    "type": "string"
                },
                "hostname": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "ipAddress": {
                    "type": "string"
                },
                "osVersion": {
                    "type": "string"
                },
                "port": {
                    "type": "integer"
                },
                "room": {
                    "type": "string"
                }
            }
        },
...
        "store.Agent": {
            "type": "object",
            "properties": {
                "agentVersion": {
                    "type": "string"
                },
                "building": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "hostname": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "ipAddress": {
                    "type": "string"
                },
                "labId": {
                    "type": "string"
                },
                "lastSeen": {
                    "type": "string"
                },
                "osVersion": {
                    "type": "string"
                },
                "pendingUpdate": {
                    "type": "string"
                },
                "port": {
                    "type": "integer"
                },
                "room": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updateOfferedAt": {
                    "description": "Staggered-rollout state. UpdateOfferedAt is nil when the agent is not\ncurrently in flight for an auto-update; UpdateTargetVersion is the version\nit was last offered toward.",
                    "type": "string"
                },
                "updateTargetVersion": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
    }
}